requirements.txt
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
ddt==1.2.1
diff-match-patch==20121119
Django==2.0.5
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.4
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.1
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
PyMySQL==0.8.1
pyparsing==2.4.2
pytesseract==0.3.0
python-dateutil==2.8.0
pytz==2018.4
PyYAML==3.12
requests==2.18.4
selenium==3.141.0
six==1.11.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
